Don't allow empty fast5/fastq fail/pass folder
As a failsafe and part of our business rule at QBiC,
we must ensure that the pass and fail folder for either
fast5 or fastq are not empty during the creation of
an OxfordNanoporeMeasurement object.

Therefore, if the pass and fail folder are both empty for either
fastq oder fast5, an IllegalStateException will be thrown
during the instanciation of an OxfordNanopreMeasurement object.